+++ packages/reactivity-core/utils/reportTaskError.ts
@@ -1,6 +1,12 @@
+let globalReportError: typeof globalThis["reportError"];
+if (typeof globalThis.reportError === "function") {
+    globalReportError = globalThis.reportError;
+} else {
+    globalReportError = (err: Error) => console.error(err);
+}
+
+
 // This function is being mocked in tests
 export function reportTaskError(e: unknown) {
-    // This makes the error an unhandled rejection for lack of a better
-    // reporting mechanism. Stupid idea?
-    Promise.reject(new Error(`Error in effect or watch callback`, { cause: e }));
+    globalReportError(new Error("Error in effect or watch callback", { cause: e }));
 }