+++ packages/reactivity-core/CHANGELOG.md
@@ -36,6 +36,8 @@
     );
     ```
 
+-   Use the browser's [`reportError`](https://developer.mozilla.org/en-US/docs/Web/API/Window/reportError) function (when available) to report errors from async callback executions.
+
 ## v0.4.3
 
 -   Introduce `synchronized`, a new kind of signal designed to integrate foreign data sources.